+++ examples/sandbox/sandbox.py
@@ -14,6 +14,11 @@ Contributors: salvadordura@gmail.com
 #
 ###############################################################################
 
+"""
+tut_artif.py 
+
+Tutorial on artificial cells (no sections)
+"""
 
 from netpyne import specs, sim
 from netpyne.specs import Dict
@@ -27,13 +32,19 @@ simConfig = specs.SimConfig()  # dictionary to store sets of simulation configur
 ###############################################################################
 
 # Population parameters
-netParams.popParams['PYR1'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 1} # add dict with params for this pop 
-netParams.popParams['PYR2'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 1} # add dict with params for this pop 
-netParams.popParams['background'] = {'cellModel': 'NetStim', 'numCells': 2, 'rate': 20, 'noise': 0.5, 'start': 1, 'seed': 2}  # background inputs
+#netParams.popParams['PYR1'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 10} # pop of HH cells
+#netParams.popParams['artifVec'] = {'cellModel': 'VecStim', 'numCells': 10000, 'interval': 100, 'noise': 0.5, 'start': 50}  # pop of NetStims
+netParams.popParams['artifNet'] = {'cellModel': 'VecStim', 'numCells': 1000, 'rate': 5, 'noise': 0.5, 'start': 50, 
+    'pulses': [{'start': 200, 'end': 300, 'rate': 50, 'noise':0.2}, {'start': 500, 'end': 800, 'rate': 30, 'noise':0.5}]}  # pop of NetStims
 
 # Synaptic mechanism parameters
-netParams.addSynMechParams('AMPA', {'mod': 'Exp2Syn', 'tau1': 0.1, 'tau2': 1.0, 'e': 0})
-netParams.addSynMechParams('esyn', {'mod': 'ElectSyn', 'g': 0.000049999999999999996})
+netParams.synMechParams['AMPA'] = {'mod': 'Exp2Syn', 'tau1': 0.1, 'tau2': 1.0, 'e': 0}
+
+
+# Stimulation parameters
+# netParams.stimSourceParams['background'] = {'type': 'NetStim', 'interval': 100, 'number': 1e5, 'start': 500, 'noise': 0.5}  # stim using NetStims after 500ms
+# netParams.stimTargetParams['bkg->PYR1'] = {'source': 'background', 'conds': {'popLabel': 'PYR1'}, 'sec':'soma', 'loc': 0.5, 'weight': 0.5, 'delay': 1}
+
 
 # Cell parameters
 ## PYR cell properties
@@ -43,22 +54,26 @@ cellParams.secs.soma.mechs.hh = {'gnabar': 0.12, 'gkbar': 0.036, 'gl': 0.003, 'e
 cellParams.conds = {'cellType': 'PYR'}
 netParams.cellParams['PYR'] = cellParams
 
-# Connections
-netParams.connParams['bg->PYR1'] = {
-    'preConds': {'popLabel': 'background'}, 'postConds': {'popLabel': 'PYR1'}, # background -> PYR
-    'weight': 0.1,                    # fixed weight of 0.08
-    'synMech': 'AMPA',                     # target NMDA synapse
-    'delay': 'uniform(1,5)'}           # uniformly distributed delays between 1-5ms
 
-netParams.connParams['PYR1->PYR2'] = {
-    'preConds': {'popLabel': 'PYR1'}, 'postConds': {'popLabel': 'PYR2'}, # PYR1 -> PYR2 (gap junction)
-    'weight': 200.0,              
-    'synMech': 'esyn',                   
-    'gapJunction': True,
-    'sec': 'soma',
-    'loc': 0.5,
-    'preSec': 'soma',
-    'preLoc': 0.5}        
+# Connections
+netParams.connParams['artif1->PYR1'] = {
+    'preConds': {'popLabel': 'artif1'}, 'postConds': {'popLabel': 'PYR1'},
+    'convergence': 8,
+    'weight': 0.005,                    
+    'synMech': 'AMPA',                
+    'delay': 'uniform(1,5)'}          
+
+# netParams.connParams['PYR1->artif2'] = {
+#     'preConds': {'popLabel': 'PYR1'}, 'postConds': {'popLabel': 'artif2'},
+#     'probability': 0.2,
+#     'weight': 0.2,                     
+#     'delay': 'uniform(1,5)'}     
+
+# netParams.addConnParams('artif2->artif3',
+#     {'preConds': {'popLabel': 'artif2'}, 'postConds': {'popLabel': 'artif3'}, 
+#     'divergence': 20,
+#     'weight': 0.05,              
+#     'delay': 3})        
 
 
 ###############################################################################
@@ -84,6 +99,5 @@ simConfig.analysis['plotRaster'] = True
 ###############################################################################
 
 sim.createSimulateAnalyze()
-